Add files via upload
These data inputs included here are updated to 2022 and intended to be used in optimization for 2023-2025. Note that base daily demand quantity and price are estimated based on reference case projected by the U.S. EIA's Annual Energy Outlook 2023. Users may want to update data inputs to solve for other periods of interest.